refactor emails and letters
