+++ src/admin/containers/add_member.js
@@ -1,39 +1,34 @@
-'use strict'
-
 const React = require('react')
 const { connect } = require('react-redux')
 const { reduxForm } = require('redux-form')
-const { post } = require('app/http')
 
 const MemberFields = require('../dumb_components/member_fields.js')
 const { fields, validate, required } = require('../form_fields/member.js')
 const { create_member } = require('../redux/modules/member.js')
+const { send_welcome } = require('../redux/modules/email.js')
 
 const buttons = (
-  { fields: { id: { value: id }, primary_email: { value: primary } }, error }
+  { fields: { id, primary_email }, error, email_handler }
 ) =>
   <div>
     { error ? <div className='error'>{error}</div> : '' }
-    { id
+    { id.value
       ? <div>
-          { letter_or_email(primary, id) }
-          <div id='member-num'>Member ID is: {id} </div>
+          { letter_or_email(email_handler, primary_email.value, id.value) }
+          <div id='member-num'>Member ID is: {id.value} </div>
         </div>
       : <button type='submit'>Submit</button>
     }
   </div>
 
-const letter_or_email = (primary_email, id) =>
-  primary_email
-  ? <a onClick={sendEmail(primary_email)} href='#/'>
+const letter_or_email = (email_handler, email, id) =>
+  email
+  ? <button onClick={() => email_handler(email)} type="button">
       Send Welcome Email
-    </a>
-  : <a href={'#/email/' + id}>
+    </button>
+  : <a href={'#/letter/' + id}>
       Print Welcome Letter
     </a>
-
-const sendEmail = primary_email =>
-  () => post({ email: primary_email}, '/api/members/welcome').fork(console.log.bind(console), console.log.bind(console))
 
 const AddMember = reduxForm(
   { form: 'member'
@@ -44,6 +39,7 @@ const AddMember = reduxForm(
 
 const NewMember = (
   { create_member
+  , send_welcome
   }
 ) => (
   <div>
@@ -52,6 +48,7 @@ const NewMember = (
       <AddMember
         fields={fields}
         Buttons={buttons}
+        button_props={ { email_handler: send_welcome } }
         onSubmit={create_member}
         required={required}
         mode='edit'
@@ -66,6 +63,7 @@ const NewMember = (
 const map_state_to_props = () => ({})
 const map_dispatch_to_props = (
   { create_member
+  , send_welcome
   }
 )
