+++ src/admin/redux/modules/email.js
@@ -1,34 +1,9 @@
-import { FETCHED_MEMBER } from './member.js'
-import { mergeAll, converge, unapply, compose, objOf, prop } from 'ramda'
+const { createAction, handleAction } = require('redux-actions')
+const { post } = require('app/http')
 
-export default (state = { id: 0, address: [], name: '' }, { type, payload }) => {
-  switch (type) {
-    case FETCHED_MEMBER:
-      return converge
-        ( unapply(mergeAll)
-        , [ compose(objOf('name'), getName)
-          , compose(objOf('address'), getAddress)
-          , compose(objOf('id'), prop('id'))
-          ]
-        )(payload)
-    default:
-      return state
-  }
-}
+const SEND_WELCOME =
+  'SEND_WELCOME'
 
-const getName = ({ title, full_name }) => title + ' ' + full_name
+export const send_welcome =
+  createAction( SEND_WELCOME, email => post({ email }, '/api/members/welcome'))
 
-const getAddress = (
-  { title
-  , first_name = ''
-  , last_name = ''
-  , address1
-  , address2
-  , address3
-  , address4
-  , county
-  , postcode
-  }
-) =>
-  [ `${title} ${first_name[0] || ''} ${last_name}`, address1, address2,
-    address3, address4, county, postcode ]
