+++ src/styles/pages/_letter.scss
@@ -1,4 +1,4 @@
-.email-page {
+.letter-page {
   width: 80%;
   .letter, .letter * {
     font-size: larger;
